NextJS: Render SignedAppUrl
