+++ web/nextjs/src/app/dashboard/account/layout.tsx
@@ -1,0 +1,9 @@
+import SharedLayout from "@/components/SharedLayout";
+
+export default function AccountLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <SharedLayout>{children}</SharedLayout>;
+}
