+++ web/nextjs/src/app/dashboard/chat/page.tsx
@@ -1,6 +1,13 @@
+import { getUser } from "@/lib/mocks";
+import { getCounselSignedAppUrl } from "@/lib/server";
+import ChatPage from "@/components/ChatPage";
+
 /**
  * Chat page is the most important for showcasing the embedded app as this is where its hosted.
  */
-export default function ChatPage() {
-  return <div>Chat</div>;
+export default async function Chat() {
+  const user = getUser();
+  const signedAppUrl = await getCounselSignedAppUrl(user.id);
+
+  return <ChatPage signedAppUrl={signedAppUrl} />;
 }
