+++ web/nextjs/src/app/dashboard/orders/layout.tsx
@@ -1,0 +1,9 @@
+import SharedLayout from "@/components/SharedLayout";
+
+export default function OrdersLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <SharedLayout>{children}</SharedLayout>;
+}
