+++ web/nextjs/src/app/dashboard/orders/page.tsx
@@ -1,8 +1,7 @@
-
 import { OrderPage } from "@/components/OrderPage";
 import { getAllOrders } from "@/lib/mocks";
 
-export default function OrdersPage() {
+export default function Orders() {
   const orders = getAllOrders();
   return <OrderPage orders={orders} />;
 }
