+++ web/nextjs/src/app/dashboard/page.tsx
@@ -1,7 +1,11 @@
 import DashboardPage from "@/components/DashboardPage";
 import { getRecentOrders } from "@/lib/mocks";
-
+import SharedLayout from "@/components/SharedLayout";
 export default function Dashboard() {
   const recentOrders = getRecentOrders();
-  return <DashboardPage recentOrders={recentOrders} />;
+  return (
+    <SharedLayout>
+      <DashboardPage recentOrders={recentOrders} />
+    </SharedLayout>
+  );
 }
