+++ web/nextjs/src/components/SharedLayout.tsx
@@ -1,0 +1,7 @@
+export default function SharedLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <div className="p-6 mx-auto container">{children}</div>;
+}
